Refactor runtime scripts into a subdirectory
