+++ src/runtime-scripts/initializeSearch.ts
